# openapi.yaml
openapi: 3.0.3
info:
  title: Expense Splitter API
  description: >
    Backend API for the Expense Splitter household ledger.
    Tracks members, one-off expenses split by fixed amounts, and pairwise
    settlements between household members.


    Authentication: **none**. Per the product spec ("no password — pick your
    identity for now"), the app has no login flow. Every endpoint below is
    unauthenticated and may be called by anyone with network access to the
    API. The "current person" the UI operates as is just a locally-selected
    member id passed as a path parameter or request field — it is an
    identity selection, not a credential, and the API does not verify it
    against any session.
  version: "1.0.0"
servers:
  - url: /api
...
paths:
...
  /members/{personId}/balances:
    get:
      tags: [members]
      summary: Get one member's pairwise balances
      description: >
        Returns, for the given person, one row per other household member
        describing the net pairwise balance between them. A positive
        `amount` means the other person owes `personId`; a negative amount
        means `personId` owes the other person. Balances are derived from
        all expenses (splits) and settlements recorded so far and are never
        simplified across more than two people. No authentication required
        — `personId` is simply which member's ledger view to compute.
      operationId: getMemberBalances
      security: []
      parameters:
        - name: personId
          in: path
          required: true
          description: Id of the member to compute balances for.
          schema:
            type: string
          example: m1
      responses:
        "200":
          description: Pairwise balances for the given member.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BalanceRow"
              example:
                - person:
                    id: m2
                    name: Tobias Renn
                    short: Tobias
                    initials: TR
                  amount: 12.50
                - person:
                    id: m3
                    name: Maren Holt
                    short: Maren
                    initials: MH
                  amount: -8.00
        "404":
          description: No member exists with the given id.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  schemas:
    Member:
      type: object
      required: [id, name, short, initials]
      properties:
        id:
          type: string
          description: Unique member id.
          example: m1
        name:
          type: string
          description: Full display name.
          example: Iris Calloway
        short:
          type: string
          description: Short display name used in compact UI.
          example: Iris
        initials:
          type: string
          description: Two-letter initials used for the avatar badge.
          example: IC
...
    BalanceRow:
      type: object
      required: [person, amount]
      properties:
        person:
          $ref: "#/components/schemas/Member"
        amount:
          type: number
          format: double
          description: >
            Net balance between the requested member and `person`. Positive
            means `person` owes the requested member; negative means the
            requested member owes `person`; zero means settled up.
          example: 12.50
...
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: Nothing outstanding between these two.